% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACToR.parse.fillin.R
\name{ACToR.parse.fillin}
\alias{ACToR.parse.fillin}
\title{Parse the ACToR files}
\usage{
ACToR.parse.fillin(
  datasets = c("Denmark", "Germany", "HealthCanada", "Japan", "Michigan",
    "Norway_Nordic", "Ohio", "Sweden")
)
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Parse the ACToR files
}
