% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACToR.parse.R
\name{ACToR.parse}
\alias{ACToR.parse}
\title{Parse the ACToR files and fill in the remaining missed data}
\usage{
ACToR.parse(dataset = "PFAS_3M")
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Parse the ACToR files and fill in the remaining missed data
}
