% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHANES_extract.R
\name{extract_NHANES}
\alias{extract_NHANES}
\title{Code to format the NHANES data
Note that this will only run under R 4.x}
\usage{
extract_NHANES(dir = "../data/", dataset = "PFAS_J")
}
\arguments{
\item{dir}{The directory to look for data}

\item{dataset}{The name of the NHANES dataset}
}
\description{
Code to format the NHANES data
Note that this will only run under R 4.x
}
