% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasBiomonitoringMerge.R
\name{pfasBiomonitoringMerge}
\alias{pfasBiomonitoringMerge}
\title{Merge the ACToR data with NHANES and put in a common format}
\usage{
pfasBiomonitoringMerge()
}
\description{
Merge the ACToR data with NHANES and put in a common format
}
