% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasBloodLevelxChainLength.R
\name{pfasBloodLevelxChainLength}
\alias{pfasBloodLevelxChainLength}
\title{Plotthe blood levels by chain length}
\usage{
pfasBloodLevelxChainLength(to.file = F)
}
\arguments{
\item{to.file}{If TRUE, write graphs to a file}

\item{data.version}{Label of folder where input data sits}
}
\description{
Plotthe blood levels by chain length
}
