% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasChemicalTable.R
\name{pfasChemicalTable}
\alias{pfasChemicalTable}
\title{Generate a table of the chemicals and the data we have for them}
\usage{
pfasChemicalTable()
}
\description{
Generate a table of the chemicals and the data we have for them
}
