% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasHalfLife.R
\name{pfasHalfLife}
\alias{pfasHalfLife}
\title{Extract the half-life information from Dawson et al. supplemental file}
\usage{
pfasHalfLife(species = "Human", route = "Oral")
}
\arguments{
\item{species}{The species to use}

\item{route}{The exposre route}
}
\description{
Extract the half-life information from Dawson et al. supplemental file
}
