% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasInVitroVsRat.R
\name{pfasInVitroVsRat}
\alias{pfasInVitroVsRat}
\title{Compare the in vitro PODs vs the Rat POD concentration}
\usage{
pfasInVitroVsRat(to.file = F)
}
\arguments{
\item{to.file}{If TRUE, write graphs to a file}

\item{data.version}{Label of folder where input data sits}
}
\description{
Compare the in vitro PODs vs the Rat POD concentration
}
