% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasMoeBoxplot.R
\name{pfasMoeBoxplot}
\alias{pfasMoeBoxplot}
\title{Plot the distribution of PODs}
\usage{
pfasMoeBoxplot(to.file = F)
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Plot the distribution of PODs
}
