% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasMoeBoxplotOPPT.R
\name{pfasMoeBoxplotOPPT}
\alias{pfasMoeBoxplotOPPT}
\title{Plot the distribution of PODs}
\usage{
pfasMoeBoxplotOPPT(to.file = F)
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Plot the distribution of PODs
}
