% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasPartitionCoefficients.R
\name{pfasPartitionCoefficients}
\alias{pfasPartitionCoefficients}
\title{Get the partition coefficients}
\usage{
pfasPartitionCoefficients(species = "Human")
}
\arguments{
\item{to.file}{If TRUE, write graphs to a file}

\item{data.version}{Label of folder where input data sits}
}
\description{
Get the partition coefficients
}
